🎨 :: router 경로 통일
